73 mjliggr fr anvndaren skapa ny aktivitet

+++ backend/src/main/java/com/kollicon/controller/ScheduleController.java
@@ -4,7 +4,10 @@ import com.kollicon.model.ScheduleModel;
 import com.kollicon.service.ScheduleService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.Optional;
 
@@ -13,8 +16,12 @@ import java.util.Optional;
 @RequestMapping("/api")
 public class ScheduleController {
 
-    @Autowired
     private ScheduleService scheduleService;
+
+    @Autowired
+    public ScheduleController (ScheduleService scheduleService) {
+        this.scheduleService = scheduleService;
+    }
 
     @PostMapping("/schedule/post")
     public ScheduleModel createSchedule(@RequestBody @Valid ScheduleModel schedule) {
@@ -23,21 +30,34 @@ public class ScheduleController {
 
     @GetMapping("/schedule/get/{id}")
     public Optional<ScheduleModel> getScheduleById(@PathVariable Long id){
-        return scheduleService.getScheduleById(id);
+        if (scheduleService.getScheduleById(id).isEmpty()) {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, "Schedule with id " + id + " not found");
+        } else {
+            return scheduleService.getScheduleById(id);
+        }
+
     }
 
     @PutMapping("/schedule/update")
-    public Optional<ScheduleModel> updateSchedule(@Valid @RequestBody ScheduleModel scheduleModel){
-        return scheduleService.updateSchedule(scheduleModel);
+    public Optional<ScheduleModel> updateSchedule(@Valid @RequestBody ScheduleModel scheduleModel) {
+        if (scheduleService.getScheduleById(scheduleModel.getId()).isEmpty()) {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, "Schedule with id " + scheduleModel.getId() + " not found");
+        } else {
+            return scheduleService.updateSchedule(scheduleModel);
+        }
     }
 
     @DeleteMapping("/schedule/delete/{id}")
     public String deleteSchedule(@PathVariable Long id) {
-        scheduleService.deleteSchedule(id);
-        return "Schedule was deleted";
+        if (scheduleService.getScheduleById(id).isEmpty()) {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, "Schedule with id " + id + " not found");
+        } else {
+            scheduleService.deleteSchedule(id);
+            return "Schedule was deleted";
         }
-
-
-
+    }
 
 }
